bundles: Resolve content chroots with full chroot
Files in the content chroots specified by the "contents" keyword in the
bundle definition file are compared against the full chroot to resolve
conflicts which are fatal. When there are no conflicts, files from the
content chroots are copied into the full chroot and added to the
corresponding bundle-info file.

Fixes #705

Signed-off-by: John Akre <john.w.akre@intel.com>